refactor(tests): Use common test utils functions
