+++ backend/pyproject.toml
@@ -12,6 +12,8 @@ package-mode = false
 
 [tool.poetry.dependencies]
 python = "^3.13"
+passlib = "^1.7.4"
+pyjwt = "^2.10.1"
 
 [tool.poetry.group.dev.dependencies]
 pre-commit = "^2.20.0"
